+++ docs/build
@@ -1,4 +1,4 @@
 #!/bin/bash -eu
 
-sphinx-apidoc -efT -o docs/_apidoc mao_merge_45m
+sphinx-apidoc -efT -o docs/_apidoc mao_45m_merge
 sphinx-build -a docs docs/_build
