+++ mao_45m_merge/power_meter.py
