+++ tests/test_metadata.py
@@ -1,5 +1,5 @@
 # third-party packages
-import mao_merge_45m
+import mao_45m_merge
 
 
 # constants
@@ -9,8 +9,8 @@ VERSION: str = "0.2.1"
 
 # test functions
 def test_author() -> None:
-    assert mao_merge_45m.__author__ == AUTHOR
+    assert mao_45m_merge.__author__ == AUTHOR
 
 
 def test_version() -> None:
-    assert mao_merge_45m.__version__ == VERSION
+    assert mao_45m_merge.__version__ == VERSION
